The simulation base for the dual module simulation is constructed. The simulations take too much time and the sample rate required to be 0.5usec for some reasons that are deeply investigated.
If the simulation is run with the scopes, the simulation length almost doubles and the ram usage increases around 1GB (or more).

To overcome this problem, the simulation results are exported to workspace, that way the simulation speed is increased signifcantly and the ram usage is decreased. (a piece of advice, use export to workspace)